requirements.txt
msgspec~=0.18.6
PyYAML~=6.0.2
pandas~=2.2.3
tqdm~=4.67.0
instructor~=1.6.4
ollama~=0.3.3
openai~=1.54.4
Django~=5.1.3
pydantic~=2.9.2
transformers~=4.46.2
accelerate~=1.1.1
streamlit~=1.40.1
icecream
groq